Split of ItemsList
