+++ src/components/dashboard/containers/Lists/theme.css
@@ -1,0 +1,55 @@
+.card {
+    display: inline-block !important;
+    margin: 10px;
+    width: 270px !important;
+}
+
+.cardTitle {
+    overflow: hidden;
+    & > div {
+        overflow: hidden;
+        & h5 {
+            width: 100%;
+            overflow: hidden;
+            white-space: nowrap;
+            text-overflow: ellipsis;
+        }
+    }
+}
+
+.cardTitle, 
+.cardMedia {
+    cursor: pointer;
+}
+
+.cardActions {
+    justify-content: space-between !important;
+}
+
+.rowCell {
+    height: 100px !important;
+}
+
+.link{
+    cursor: pointer !important;
+}
+
+
+.img {
+    width: auto;
+    height: 100%;
+    max-width: 180px;
+}
+
+.content {
+    & > img {
+        position: absolute;
+        top: 0;
+        bottom: 0;
+        left: 0;
+        right: 0;
+        margin: auto;
+        min-width: 50%;
+        min-height: 50%;
+    }
+}
